[#118] redesign actor
- refactoring property and its test
