+++ dss-core/src/main/java/io/github/ztkmkoo/dss/core/actor/property/DssNetworkActorProperty.java
@@ -8,17 +8,11 @@ public interface DssNetworkActorProperty {
 
     /**
      * netty boss event loop thread count
-     * If need override this
      */
-    default int getBossThreadCount() {
-        return 1;
-    }
+    int getBossThreadCount();
 
     /**
      * netty worker event loop thread count
-     * If need override this
      */
-    default int getWorkerThreadCount() {
-        return 0;
-    }
+    int getWorkerThreadCount();
 }
