+++ dss-core/src/test/java/io/github/ztkmkoo/dss/core/actor/property/DssNetworkActorPropertyTest.java
@@ -11,23 +11,13 @@ import static org.junit.jupiter.api.Assertions.*;
 class DssNetworkActorPropertyTest {
 
     @Test
-    void normal() {
-        final TestProperty1 property = new TestProperty1();
-        assertEquals(1, property.getBossThreadCount());
-        assertEquals(0, property.getWorkerThreadCount());
-    }
-
-    @Test
-    void overrideMethod() {
-        final TestProperty2 property = new TestProperty2();
+    void testAll() {
+        final TestProperty property = new TestProperty();
         assertEquals(2, property.getBossThreadCount());
         assertEquals(16, property.getWorkerThreadCount());
     }
 
-    private static class TestProperty1 implements DssNetworkActorProperty {
-    }
-
-    private static class TestProperty2 implements DssNetworkActorProperty {
+    private static class TestProperty implements DssNetworkActorProperty {
         @Override
         public int getBossThreadCount() {
             return 2;
